DS-160: refactoring
