+++ ngx-fudis/projects/ngx-fudis/src/lib/directives/tooltip/tooltip-api.directive.ts
@@ -1,6 +1,6 @@
 import { Directive, Input } from '@angular/core';
 import { MatTooltip } from '@angular/material/tooltip';
-import { FudisTooltipPosition } from '../../types/typography';
+import { FudisTooltipPosition } from '../../types/miscellaneous';
 
 @Directive({
 	selector: '[fudisTooltipApi]',
