--- ngx-fudis/projects/ngx-fudis/src/lib/types/expandables.ts
@@ -1,3 +1,0 @@
-export enum ExpandableType {
-	regular = 'regular',
-}
