--- ngx-fudis/projects/ngx-fudis/src/lib/types/lists.ts
@@ -1,5 +1,0 @@
-export interface FudisDescriptionListItem {
-	key: string;
-	subHeading?: string;
-	value: string;
-}
