+++ ngx-fudis/projects/ngx-fudis/src/lib/types/miscellaneous.ts
@@ -1,0 +1,13 @@
+export type FudisBadgeVariant = 'accent' | 'danger' | 'primary' | 'secondary' | 'success';
+
+export interface FudisDescriptionListItem {
+	key: string;
+	subHeading?: string;
+	value: string;
+}
+
+export type FudisExpandableType = 'regular';
+
+export type FudisSpacing = 'none' | 'xxs' | 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
+
+export type FudisTooltipPosition = 'left' | 'right' | 'above' | 'below';
