--- ngx-fudis/projects/ngx-fudis/src/lib/types/spacing.ts
@@ -1,1 +1,0 @@
-export type FudisSpacing = 'none' | 'xxs' | 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
