+++ ngx-fudis/projects/ngx-fudis/src/lib/types/typography.ts
@@ -3,5 +3,3 @@ export type FudisHeadingTag = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
 export type FudisHeadingSize = 'xxs' | 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
 
 export type FudisBodyText = 'lg-regular' | 'md-regular' | 'sm-regular' | 'lg-light' | 'md-light';
-
-export type FudisTooltipPosition = 'left' | 'right' | 'above' | 'below';
