+++ ngx-fudis/projects/ngx-fudis/src/lib/utilities/config.service.ts
@@ -13,21 +13,11 @@ export class FudisConfigService {
 		datepicker: { closeLabel: 'Close calendar' },
 	});
 
-	private _joo = signal<string>('moi');
-
 	/**
 	 * To set from application default values for all components application uses.
 	 */
 	setConfig(defaultValues: FudisFormConfig): void {
 		this._defaultFormConfig.set({ ...defaultValues });
-	}
-
-	setJoo(value: string) {
-		this._joo.set(value);
-	}
-
-	getJoo(): Signal<string> {
-		return this._joo.asReadonly();
 	}
 
 	/**
